evals/eval_from_golden_questions/pyproject.toml
[project]
name = "eval-from-golden-questions"
version = "0.1.0"
description = "RAG evaluation system using DeepEval metrics for Kunnskapsassistenten"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "kua-shared @ file:///Users/benjaminhope/src/customers/dio/KuA_evals/agents",
    "deepeval>=1.0.0",
    "langfuse>=3.10.0",
    "diskcache>=5.6.0",
    "tqdm>=4.66.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.2.1",
    "pymongo>=4.15.5",
    "pandas>=2.3.3",
    "streamlit>=1.52.2",
    "plotly>=6.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=term-missing --cov-report=html"
